move adapter classes into proxy file
