+++ common/src/python/flywheel_adaptor/flywheel_proxy.py
@@ -6,8 +6,9 @@ from typing import List, Mapping, Optional
 import flywheel
 from flywheel import (Client, ContainerIdViewInput, DataView, GearRule,
                       GearRuleInput, RolesRole, ViewIdOutput)
+from flywheel.models.access_permission import AccessPermission
 from flywheel.models.project_parents import ProjectParents
-from flywheel_adaptor.group_adaptor import GroupAdaptor
+from flywheel.models.roles_role_assignment import RolesRoleAssignment
 from fw_client import FWClient
 from fw_utils import AttrDict
 
@@ -71,7 +72,7 @@ class FlywheelProxy:
         """
         return self.__fw.groups.find(f'_id={group_id}')
 
-    def find_group(self, group_id: str) -> Optional[GroupAdaptor]:
+    def find_group(self, group_id: str) -> Optional['GroupAdaptor']:
         """Returns group for group id.
 
         Args:
@@ -458,3 +459,432 @@ class FlywheelProxy:
     def get_site(self):
         """Returns URL for site of this instance."""
         return self.__fw.get_config()["site"]["redirect_url"]
+
+class GroupAdaptor:
+    """Defines an adaptor for a flywheel group."""
+
+    def __init__(self, *, group: flywheel.Group, proxy: FlywheelProxy) -> None:
+        self.__group = group
+        self.__fw = proxy
+
+    @property
+    def label(self) -> str:
+        """Return the label of the group."""
+        return self.__group.label
+
+    def proxy(self) -> FlywheelProxy:
+        """Return the proxy for the flywheel instance."""
+        return self.__fw
+
+    def projects(self) -> List[flywheel.Project]:
+        """Return projects for the group."""
+        return self.__group.projects()
+
+    def get_tags(self) -> List[str]:
+        """Return the list of tags for the group.
+
+        Returns:
+          list of tags for the group
+        """
+        return self.__group.tags
+
+    def add_tag(self, tag: str) -> None:
+        """Adds the tag to the group for the center.
+
+        Args:
+          tag: the tag to add
+        """
+        self.__group.add_tag(tag)
+
+    def get_group_users(self,
+                        *,
+                        access: Optional[str] = None) -> List[flywheel.User]:
+        """Gets the users for the named group.
+
+        Returns an empty list if the group does not exist or there are no
+        user roles.
+        If a role is specified, only the users with the role will be returned.
+
+        Args:
+          group_name: the group ID
+          role: (optional) the role id
+        Returns:
+          the list of users for the group
+        """
+        permissions = self.__group.permissions
+        if not permissions:
+            return []
+
+        if access:
+            permissions = [
+                permission for permission in permissions
+                if access == permission.access
+            ]
+
+        user_ids = [
+            permission.id for permission in permissions if permission.id
+        ]
+        users = []
+        for user_id in user_ids:
+            user = self.__fw.find_user(user_id)
+            if user:
+                users.append(user)
+        return users
+
+    def get_user_access(self) -> List[AccessPermission]:
+        """Returns the user access for the group.
+
+        Returns:
+          the access permissions for the group
+        """
+        return self.__group.permissions
+
+    def add_user_access(self, new_permission: AccessPermission) -> None:
+        """Adds permission for user to access the group of the center.
+
+        Args:
+          permission: permission object indicating user and group access
+        """
+        if not new_permission.id:
+            log.error('new permission has no user ID to add to group %s',
+                      self.__group.label)
+            return
+
+        if not new_permission.access:
+            log.warning('new permission for user %s has no access, skipping',
+                        new_permission.id)
+            return
+
+        if self.__fw.dry_run:
+            log.info('Dry Run: would add access %s for user %s to group %s',
+                     new_permission.access, new_permission.id,
+                     self.__group.label)
+            return
+
+        existing_permissions = [
+            perm for perm in self.__group.permissions
+            if perm.id == new_permission.id
+        ]
+        if not existing_permissions:
+            self.__group.add_permission(new_permission)
+            return
+
+        self.__group.update_permission(
+            new_permission.id,
+            AccessPermission(id=None, access=new_permission.access))
+
+    def add_role(self, new_role: RolesRole) -> None:
+        """Add the role to the the group for center.
+
+        Args:
+          new_role: the role to add
+        """
+        if not self.__fw:
+            log.error('no Flywheel proxy given when adding users to group %s',
+                      self.__group.label)
+            return
+
+        self.__fw.add_group_role(group=self.__group, role=new_role)
+
+    # TODO: should return ProjectAdaptor
+    def get_project(self, label: str) -> Optional[flywheel.Project]:
+        """Returns a project in this group with the given label.
+
+        Creates a new project if none exists.
+
+        Args:
+          label: the label for the project
+        Returns:
+          the project in this group with the label
+        """
+        return self.__fw.get_project(group=self.__group, project_label=label)
+
+    def find_project(self, label: str) -> Optional['ProjectAdaptor']:
+        """Returns the project adaptor in the group with the label.
+
+        Args:
+          label: the label of the desired project
+        Returns:
+          Project adaptor for project with label if exists, None otherwise.
+        """
+        projects = self.__fw.find_projects(group_id=self.__group.id,
+                                           project_label=label)
+        if not projects:
+            return None
+
+        return ProjectAdaptor(project=projects[0], proxy=self.__fw)
+
+class ProjectAdaptor:
+    """Defines an adaptor for a flywheel project."""
+
+    def __init__(self, *, project: flywheel.Project,
+                 proxy: FlywheelProxy) -> None:
+        self.__project = project
+        self.__fw = proxy
+
+    def __pull_project(self) -> None:
+        """Pulls the referenced project from Flywheel instance."""
+        projects = self.__fw.find_projects(group_id=self.group,
+                                           project_label=self.label)
+        if not projects:
+            return
+
+        self.__project = projects[0]
+
+    # pylint: disable=(invalid-name)
+    @property
+    def id(self):
+        """Returns the ID of the enclosed project."""
+        return self.__project.id
+
+    @property
+    def label(self):
+        """Returns the label of the enclosed project."""
+        return self.__project.label
+
+    @property
+    def group(self) -> str:
+        """Returns the group label of the enclosed project."""
+        return self.__project.group
+
+    def add_tag(self, tag: str) -> None:
+        """Add tag to the enclosed project.
+
+        Args:
+          tag: the tag
+        """
+        if tag not in self.__project.tags:
+            self.__project.add_tag(tag)
+
+    def set_copyable(self, state: bool) -> None:
+        """Sets the copyable state of the project to the value.
+
+        Args:
+          state: the copyable state to set
+        """
+        self.__project.update(copyable=state)
+
+    def set_description(self, description: str) -> None:
+        """Sets the description of the project.
+
+        Args:
+          description: the project description
+        """
+        self.__project.update(description=description)
+
+    def get_file(self, name: str):
+        """Gets the file from the enclosed project.
+
+        Args:
+          name: the file name
+        Returns:
+          the named file
+        """
+        return self.__project.get_file(name)
+
+    def read_file(self, name: str) -> bytes:
+        """Reads file from the named file.
+
+        Args:
+          name: the file name
+        Returns:
+          the bytes from the file
+        """
+        return self.__project.read_file(name)
+
+    def upload_file(self, file_spec: flywheel.FileSpec) -> None:
+        """Uploads the indicated file to enclosed project.
+
+        Args:
+          file_spec: the file specification
+        """
+        self.__project.upload_file(file_spec)
+
+    def get_user_roles(self, user_id: str) -> List[str]:
+        """Gets the list of user role ids in this project.
+
+        Args:
+          user_id: the user id for the user
+        Returns:
+          list of role ids
+        """
+        assignments = [
+            assignment for assignment in self.__project.permissions
+            if assignment.id == user_id
+        ]
+        if not assignments:
+            return []
+
+        return assignments[0].role_ids
+
+    def add_user_roles(self, role_assignment: RolesRoleAssignment) -> bool:
+        """Adds role assignment to the project.
+
+        Args:
+          role_assignment: the role assignment
+        Returns:
+          True if role is new, False otherwise
+        """
+        user_roles = self.get_user_roles(role_assignment.id)
+        if not user_roles:
+            log_message = (f"User {role_assignment.id}"
+                           " has no permissions for "
+                           f"project {self.__project.label}"
+                           ", adding roles")
+            if self.__fw.dry_run:
+                log.info("Dry Run: %s", log_message)
+                return True
+
+            log.info(log_message)
+            user_role = RolesRoleAssignment(id=role_assignment.id,
+                                            role_ids=role_assignment.role_ids)
+            self.__project.add_permission(user_role)
+            self.__pull_project()
+            return True
+
+        different = False
+        for role_id in role_assignment.role_ids:
+            if role_id not in user_roles:
+                different = True
+                user_roles.append(role_id)
+        if not different:
+            return False
+
+        log_message = f"Adding roles to user {role_assignment.id}"
+        if self.__fw.dry_run:
+            log.info("Dry Run: %s", log_message)
+            return True
+
+        self.__project.update_permission(
+            role_assignment.id,
+            RolesRoleAssignment(id=None, role_ids=user_roles))
+        self.__pull_project()
+        return True
+
+    def add_admin_users(self, permissions: List[AccessPermission]) -> None:
+        """Adds the users with admin access in the given group permissions.
+
+        Args:
+          permissions: the group access permissions
+        """
+        admin_role = self.__fw.get_admin_role()
+        assert admin_role
+        for permission in permissions:
+            self.add_user_roles(
+                RolesRoleAssignment(id=permission.id,
+                                    role_ids=[admin_role.id]))
+
+    def get_gear_rules(self) -> List[GearRule]:
+        """Gets the gear rules for this project.
+
+        Returns:
+          the list of gear rules
+        """
+        return self.__fw.get_project_gear_rules(project=self.__project)
+
+    def add_gear_rule(self, *, rule_input: GearRuleInput) -> None:
+        """Adds the gear rule to the Flywheel project.
+
+        Replaces an existing rule with the same name.
+
+        Args:
+          rule_input: the GearRuleInput for the gear
+        """
+        project_rules = self.__fw.get_project_gear_rules(self.__project)
+        conflict = None
+        for rule in project_rules:
+            if rule.name == rule_input.name:
+                conflict = rule
+                break
+
+        if self.__fw.dry_run:
+            if conflict:
+                log.info(
+                    'Dry Run: would remove conflicting '
+                    'rule %s from project %s', conflict.name,
+                    self.__project.label)
+            log.info('Dry Run: would add gear rule %s to project %s',
+                     rule_input.name, self.__project.label)
+            return
+
+        if conflict:
+            self.__fw.remove_project_gear_rule(project=self.__project,
+                                               rule=conflict)
+
+        self.__fw.add_project_rule(project=self.__project,
+                                   rule_input=rule_input)
+
+    def remove_gear_rule(self, *, rule: GearRule) -> None:
+        """Removes the gear rule from the project.
+
+        Args:
+          rule: the rule to remove
+        """
+        self.__fw.remove_project_gear_rule(project=self.__project, rule=rule)
+
+    def get_apps(self) -> List[AttrDict]:
+        """Returns the list of viewer apps for the project.
+
+        Returns:
+          the viewer apps for the project
+        """
+        return self.__fw.get_project_apps(self.__project)
+
+    def set_apps(self, apps: List[AttrDict]) -> None:
+        """Sets the viewer apps for the project.
+
+        Args:
+          apps: the list of viewer apps to add
+        """
+        self.__fw.set_project_apps(project=self.__project, apps=apps)
+
+    def get_dataviews(self) -> List[DataView]:
+        """Returns the list of dataviews for the project.
+
+        Returns:
+          the dataviews in the enclosed project
+        """
+        return self.__fw.get_dataviews(self.__project)
+
+    def get_dataview(self, label: str) -> Optional[DataView]:
+        """Returns the dataview in the project with the label.
+
+        Args:
+          label: the label for the desired dataview
+        Returns:
+          the dataview with matching label, None otherwise
+        """
+        dataviews = self.get_dataviews()
+        for dataview in dataviews:
+            if label == dataview.label:
+                return dataview
+
+        return None
+
+    def add_dataview(self, dataview: DataView) -> str:
+        """Adds the dataview to the enclosed project.
+
+        Args:
+          dataview: the DataView to add
+        """
+
+        # Copy the dataview into a ContainerIdViewInput
+        # which is required to add the dataview
+        # copying all of the properties but "origin"
+        view_template = ContainerIdViewInput(
+            parent=dataview.parent,
+            label=dataview.label,
+            description=dataview.description,
+            columns=dataview.columns,
+            group_by=dataview.group_by,
+            filter=dataview.filter,
+            file_spec=dataview.file_spec,
+            include_ids=dataview.include_ids,
+            include_labels=dataview.include_labels,
+            error_column=dataview.error_column,
+            missing_data_strategy=dataview.missing_data_strategy,
+            sort=dataview.sort,
+            id=dataview.id)
+        view_id = self.__fw.add_dataview(project=self.__project,
+                                         viewinput=view_template)
+        return view_id.id
